refactor: create `DashboardLayout` component and update Pilkada 2024 page
